+++ sciencebeam_gym/tools/vocabulary/extract_embeddings_vocabulary.py
