+++ sciencebeam_gym/tools/vocabulary/filter_embeddings_vocabulary.py
